backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "sqlmodel>=0.0.25",
    "uvicorn>=0.37.0",
]
